Added tax to calculation

+++ src/core/models/OrderItem.java
@@ -1,14 +1,15 @@
 package core.models;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class OrderItem {
 	private UUID id;
 	private UUID orderId;
 	private UUID productId;
-	private float priceAtOrderTime;
+	private BigDecimal priceAtOrderTime;
 	
-	public OrderItem(UUID orderId, UUID productId, float priceAtOrderTime) {
+	public OrderItem(UUID orderId, UUID productId, BigDecimal priceAtOrderTime) {
 		this.orderId = orderId;
 		this.productId = productId;
 		this.priceAtOrderTime = priceAtOrderTime;
@@ -26,7 +27,7 @@ public class OrderItem {
 		return productId;
 	}
 
-	public float getPriceAtOrderTime() {
+	public BigDecimal getPriceAtOrderTime() {
 		return priceAtOrderTime;
 	}
